[clang][test] Add lit helper for windows paths
This adds 2 new lit helpers (only to the clang config) `%{rootsep}` and
`%{sep}`, these allow writing tests that correctly handle slashes on
Windows. In the case of debug-prefix-map.c the behavior of the test was
actually incorrect because the logic being tested checks `is_absolute`
on the remapped path, which previously, even on windows, was
`/UNLIKELY_PATH/empty/foo`. Because of this the test had to be more
lenient than the actual behavior is.

Differential Revision: https://reviews.llvm.org/D111457